pura_calc_api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
simpleeval==0.9.13
pydantic==2.11.7
mysql-connector-python==9.3.0
python-dotenv==1.0.1
